+++ includes/class-wc-cache-helper.php
@@ -2,16 +2,10 @@
 /**
  * WC_Cache_Helper class.
  *
- * @class 		WC_Cache_Helper
- * @version		2.2.0
- * @package		WooCommerce/Classes
- * @category	Class
- * @author 		WooThemes
+ * @package WooCommerce/Classes
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
+defined( 'ABSPATH' ) || exit;
 
 /**
  * WC_Cache_Helper.
@@ -26,6 +20,8 @@ class WC_Cache_Helper {
 		add_action( 'admin_notices', array( __CLASS__, 'notices' ) );
 		add_action( 'delete_version_transients', array( __CLASS__, 'delete_version_transients' ) );
 		add_action( 'wp', array( __CLASS__, 'prevent_caching' ) );
+		add_action( 'clean_term_cache', array( __CLASS__, 'clean_term_cache' ), 10, 2 );
+		add_action( 'edit_terms', array( __CLASS__, 'clean_term_cache' ), 10, 2 );
 	}
 
 	/**
@@ -93,7 +89,7 @@ class WC_Cache_Helper {
 	public static function geolocation_ajax_redirect() {
 		if ( 'geolocation_ajax' === get_option( 'woocommerce_default_customer_address' ) && ! is_checkout() && ! is_cart() && ! is_account_page() && ! is_ajax() && empty( $_POST ) ) { // WPCS: CSRF ok, input var ok.
 			$location_hash = self::geolocation_ajax_get_location_hash();
-			$current_hash  = isset( $_GET['v'] ) ? wc_clean( wp_unslash( $_GET['v'] ) ) : ''; // WPCS: sanitization ok, input var ok.
+			$current_hash  = isset( $_GET['v'] ) ? wc_clean( wp_unslash( $_GET['v'] ) ) : ''; // WPCS: sanitization ok, input var ok, CSRF ok.
 			if ( empty( $current_hash ) || $current_hash !== $location_hash ) {
 				global $wp;
 
@@ -118,14 +114,14 @@ class WC_Cache_Helper {
 	/**
 	 * Get transient version.
 	 *
-	 * When using transients with unpredictable names, e.g. those containing an md5.
+	 * When using transients with unpredictable names, e.g. those containing an md5
 	 * hash in the name, we need a way to invalidate them all at once.
 	 *
-	 * When using default WP transients we're able to do this with a DB query to.
+	 * When using default WP transients we're able to do this with a DB query to
 	 * delete transients manually.
 	 *
-	 * With external cache however, this isn't possible. Instead, this function is used.
-	 * to append a unique string (based on time()) to each transient. When transients.
+	 * With external cache however, this isn't possible. Instead, this function is used
+	 * to append a unique string (based on time()) to each transient. When transients
 	 * are invalidated, the transient version will increment and data will be regenerated.
 	 *
 	 * Raised in issue https://github.com/woocommerce/woocommerce/issues/5777.
@@ -138,12 +134,33 @@ class WC_Cache_Helper {
 	public static function get_transient_version( $group, $refresh = false ) {
 		$transient_name  = $group . '-transient-version';
 		$transient_value = get_transient( $transient_name );
+		$transient_value = strval( $transient_value ? $transient_value : '' );
 
-		if ( false === $transient_value || true === $refresh ) {
-			self::delete_version_transients( $transient_value );
-			set_transient( $transient_name, $transient_value = time() );
+		if ( '' === $transient_value || true === $refresh ) {
+			$old_transient_value = $transient_value;
+			$transient_value     = (string) time();
+
+			if ( $old_transient_value === $transient_value ) {
+				// Time did not change but transient needs flushing now.
+				self::delete_version_transients( $transient_value );
+			} else {
+				self::queue_delete_version_transients( $transient_value );
+			}
+
+			set_transient( $transient_name, $transient_value );
 		}
 		return $transient_value;
+	}
+
+	/**
+	 * Queues a cleanup event for version transients.
+	 *
+	 * @param string $version Version of the transient to remove.
+	 */
+	protected static function queue_delete_version_transients( $version = '' ) {
+		if ( ! wp_using_ext_object_cache() && ! empty( $version ) ) {
+			wp_schedule_single_event( time() + 30, 'delete_version_transients', array( $version ) );
+		}
 	}
 
 	/**
@@ -158,12 +175,17 @@ class WC_Cache_Helper {
 		if ( ! wp_using_ext_object_cache() && ! empty( $version ) ) {
 			global $wpdb;
 
-			$limit    = apply_filters( 'woocommerce_delete_version_transients_limit', 1000 );
+			$limit = apply_filters( 'woocommerce_delete_version_transients_limit', 1000 );
+
+			if ( ! $limit ) {
+				return;
+			}
+
 			$affected = $wpdb->query( $wpdb->prepare( "DELETE FROM {$wpdb->options} WHERE option_name LIKE %s ORDER BY option_id LIMIT %d;", '\_transient\_%' . $version, $limit ) ); // WPCS: cache ok, db call ok.
 
-			// If affected rows is equal to limit, there are more rows to delete. Delete in 10 secs.
+			// If affected rows is equal to limit, there are more rows to delete. Delete in 30 secs.
 			if ( $affected === $limit ) {
-				wp_schedule_single_event( time() + 10, 'delete_version_transients', array( $version ) );
+				self::queue_delete_version_transients( $version );
 			}
 		}
 	}
@@ -196,9 +218,40 @@ class WC_Cache_Helper {
 		if ( $enabled && ! in_array( '_wc_session_', $settings, true ) ) {
 			?>
 			<div class="error">
-				<p><?php echo wp_kses_post( sprintf( __( 'In order for <strong>database caching</strong> to work with WooCommerce you must add %1$s to the "Ignored Query Strings" option in <a href="%2$s">W3 Total Cache settings</a>.', 'woocommerce' ), '<code>_wc_session_</code>', esc_url( admin_url( 'admin.php?page=w3tc_dbcache' ) ) ) ); ?></p>
+				<p>
+				<?php
+				/* translators: 1: key 2: URL */
+				echo wp_kses_post( sprintf( __( 'In order for <strong>database caching</strong> to work with WooCommerce you must add %1$s to the "Ignored Query Strings" option in <a href="%2$s">W3 Total Cache settings</a>.', 'woocommerce' ), '<code>_wc_session_</code>', esc_url( admin_url( 'admin.php?page=w3tc_dbcache' ) ) ) );
+				?>
+				</p>
 			</div>
 			<?php
+		}
+	}
+
+	/**
+	 * Clean term caches added by WooCommerce.
+	 *
+	 * @since 3.3.4
+	 * @param array|int $ids Array of ids or single ID to clear cache for.
+	 * @param string    $taxonomy Taxonomy name.
+	 */
+	public static function clean_term_cache( $ids, $taxonomy ) {
+		if ( 'product_cat' === $taxonomy ) {
+			$ids = is_array( $ids ) ? $ids : array( $ids );
+
+			$clear_ids = array( 0 );
+
+			foreach ( $ids as $id ) {
+				$clear_ids[] = $id;
+				$clear_ids   = array_merge( $clear_ids, get_ancestors( $id, 'product_cat', 'taxonomy' ) );
+			}
+
+			$clear_ids = array_unique( $clear_ids );
+
+			foreach ( $clear_ids as $id ) {
+				wp_cache_delete( 'product-category-hierarchy-' . $id, 'product_cat' );
+			}
 		}
 	}
 }
